refactor: 영화 목록을 추가하는 로직을 메서드로 분리
